move addIsDuplicateException to EventLogger
